Version 0.1
This version of my ChatBot is the first release on GitHub. The program is early in a beta, and is only being uploaded now due to faulty antivirus getting false positives from Visual Studio building. All the code should be here for the moment. Fiddle around with it if you want!